R628-014: Moving the cursor should stop the completion
Adapt the filters to allow the "move to" action in the editor
when the completion window is opened. Now most of the action in
the editor will canceled the completion window.
However, "move to next/previous line" should not stop the
completion but need to be disabled when the completion is active.

Suppress unused old code in text_utils.py

(no-precommit-check)

Change-Id: I3cb4b549a087782f4e86809dfe5086355152bdb9