R711-034 Keep editor scrolled to the user-set location
Make sure the editor stays scrolled to the position
that was given by the user - otherwise this is done
as a fallaback in Connect_Expose, which can take
a very long time.

Also handle the special case when the user scrolls
the text manually while the editor is being open.

Change-Id: Ia35e8ce4da99e765a03a34169c8aa2286a6c9e98