R724-044 Avoid crash when exiting several times manual proof
We are now more careful when destroying elements of the interface
(windows etc) that may have already been deleted. It means that python
variable may very well point to nothing.

* share/plug-ins/spark2014/itp_lib.py
(kill): Use the name of elements to remove them.
(exit): Do not show the yes_no_dialog if manual proof is not started.

Change-Id: I90e9df21d5775b8131bf13683ce8a358406d5180