R722-002 manual proof: no failure when no nodes are selected
Issuing a command when no nodes are selected result in a failure. Now,
if nothing is selected we still send the command on the ROOT_NODE.

(ROOT_NODE): zero by default (shared with Why3)
(interactive_console_input): Add self.called which detect if the function
call_on_select is called during iteration selected_foreach.
If not, we send the request on the root node.

Change-Id: If481e127730372a5cc02ae715ecc913cd8732acd