R724-044 Do not ask to exit when initializing proof tree
When initializing the manual proof, the message to exit sometimes appear
because first nodes are proved but not the lasts. This commit solves the
problem by starting detecting that the root nodes are proved only after
the "initialization message" was sent by the server.

(parse_notif): Initialize is_init to false when the tree is reset.
(parse_message): Adding global is_init which is set to true when the server
send the message "Session initialized successfully" (this is sent after
sending all the nodes of the tree (initially)).
(roots_is_ok): If is_init is False then the proof is not considered done

Change-Id: I6b22082a195efd0a60fd6b7fb7a3c24b467d7547