R426-012: Use 'GNATprove' category for SPARK messages
Instead of relying on the GPS default output parser, we now create
the messages directly in the SPARK output parser in order to assign
them the category we want (i.e: 'GNATprove' instead of 'Builder Results').

This allows to always have the 'GNATprove' category in all cases: when
running GNATprove with or without the Analysys Report at the end.

Change-Id: I22273b8181f246922cb8876aa745da75776bcef0