R426-012: Allow 'SPARK/Show Report' action at any time.
The 'SPARK/Show Report' action can now be executed at any time, as long
as some GNATprove messages are present in the Locations view.

This is useful when users only want to check the report from time to time:
they will then be able to disable the preference that displays it every
time GNATprove finishes and click manually on the menu instead.

Change-Id: I7acdfa01d89d7803ba018bacc0645754496eb64a