R822-035: Refresh the Learn view each when learn items are added.
A listener interface has been added to the Learn module in order
to react to changes regarding learn items (e.g: when registering
an action that should go in the Learn view).

This ensures that the Learn view is always synchronizes with the
items that need to be displayed.

A test has been added for this.

Change-Id: I775d59418abea7e052186ae30403530e3bbff6d6