R906-008: Suppress unused mechanism
The idea was when trying to display a var in the Variables view.
Only the value and the type's description were needed to parse the
type structure. And then we only need the type's name when displaying
the value. This will be done in any case so simplify the code.

As a side effect: for an array of record we were evaluating the types
only when displaying in the Variables view so we were evaluating
the types for each of the rows. Now we only evaluate the type of the
first row.

Change-Id: I86a6bca3a63df443078e6e1f6cf2b785d4b97ac6