R906-009: Prioritize the errors over the warnings
The Log stream contains warning and error messages.
However the current parsing doesn't work if an error is detected
before executing the command. Thus the first line of the log stream
is the error message. In any other case, the first line of the log
stream is the command itself and then the errors/warnings are the
following lines.

Change-Id: I582d9d5d9e8afb38c9d4d646a5ebe47ecb681032