R426-012: Few improvements in Analysis Report
Negative numbers were displayed when performing the
'Display analysis data' action two times in a row.

This was due to the fact that we clear the messages report model before
displaying any data: thus, the messages that are being removed where
decrementing the counters that were already set to 0.

A proper sorting function has also been added to the messages report,
so that the global total row always stays on top.

Actions have also been added to expand/collapse rows selected in the
Analysis Report.

Change-Id: I4ee9f7c0c93758c0572847d27a0f554bd8adb3b0