RA24-011: Don't reuse the same CSS provider on pref changed
When the preferences change, we should not reuse the same CSS provider
in order to reload our custom CSS color variables: this has for effect
to clear all the CSS data previoulsy loaded by this provider, including
colors not linked with preferences (e.g: selection color).

Instead, we now create a new provider.

Add a debounce parameter to the 'preferences_changed' python hook
in order to react only one time when multiple preferences are
changed.

Change-Id: I274765828c8c372b11facdc8df5a11748cde2e8b