RA02-047: Don't freeze the UI when lauching the debugger
Don't wait synchronously for GDB to complete the 'file' command since
this command can be pretty long depending on the executable being
executed.

This is done by adding a 'Synchronous' parameter to the
Send_And_Get_Output functions: when False, GPS will not freeze the UI
while waiting on GDB but instead will add some output filters to
retrieve the output asynchronously.

Also, remove the 'Mode' parameter of the Set_Executable command: this
command should always be visible since it can be long.

A new 'debuggee_started' hook has also been added so that python
plugins can know when the debuggee is started/connected.

Change-Id: I6ccf69e597f6e38f64fd446f3ac854dc8e04a39f