RB12-045: Fix the Interactive_Consoles.Grab_Events subprogram
The console should only grab events if the console is realized.
On the contrary, when calling Grab_Events with Grab set to False,
we should make sure to remove the grab, even when the console is
not realized.

Add defensive code in the Keymamager_Module, to be sure that we
don't process events on widgets that are not realized yet.

An automatic test has also been added.

Change-Id: I05261e8437c417b29908f2d36e4da2bdf029024c