RB09-006: Don't call Set_Size_Request on Search view
When an optional selector is displayed in the Search view (e.g: when
selecting the 'Files...' selector), increase the view's height so that
users can directly see it if it's not visible.

To achieve that, don't use Set_Size_Request: this sets the view's
minimal size so users won't be able to reduce it. Use the Resize
subprogram instead. 

Change-Id: I16a73c0a1cff808b70d19ada3d0cca7b49077a6a