S104-021: Avoid creating editor's context twice when cursor changes
The "cursor_position_changed" is now emitted everytime the
selection_bound mark changes: that mark is moved when the selection
changes but also when the cursor changes, without selecting anything.

This allows source editor views to connect to this unique signal
to handle both events (i.e: by building a new context): changes
regarding the selection and changes regarding the cursor's position.

Change-Id: I7c927a65608812b2b26417ec2967c8ef9a7749e8