S115-024 allow the task manager with nested main loops
We are processing events in nested main loops in some cases,
for instance while waiting for a debugger prompt. For these
cases, we still want the g_sources associated with the task_manager
to run: we need this to process most actions.

This fixes an issue when launching "run" or "continue"
commands using the debugger actions, which is the case when
they are invoked from menus or toolbar buttons.

Change-Id: Ib7923efe35332f62e182e3ad15321a083d870792