R620-039: Don't scroll synchronously after reloading files
When reloading a file in an editor (i.e: after detecting changes on
disk), we reset completely the underlying source buffer, which has
for effect to perform some computations/redrawings in idle functions.

Thus, to replace the cursor at its original position and scroll
properly to it, we should perform the scrolling operation in an
idle function, to make sure that all these computations/redrawings
have been done.

An automatic test has been added.

Change-Id: Ibf29d600d09b46ca77dcc6c050e9800c347f9853