S117-028: Fix MDI speratators' color
The MDI separators use the 'background-color' CSS property instead
of the 'color' one, so make sure to specify both properties in order
to cover all the types of separators.

Use a slighly darker color for dark theme separators, so that MDI
pane separators still integrate well with the editor's background
color.

Change-Id: I46d205c33698b8da79b3fe68b3781236ad158300