S329-011: Fix some issues with LSP-based tooltips
In particular, make sure to escape the tooltips text
in all cases, event when there is no style associated
with a given token while highlighting it.

Use an horizontal box to avoid the tooltip text to be centered
when there are several tooltip handlers at the same time (e.g:
codefix + LSP hover text).

Finally, don't display 'Loading...' while waiting for the
LSP response: this can lead to graphical glitches.

Change-Id: I8ed8894c8773a1f2d1cfa9ed4c680b6bd2f527bd