S510-007: Fix gnatcheck warning dialog
A new GPS.MDI.warning_dialog method has been added to ease
the creation of warning dialogs from Python.

This method is now used in the gnatcheck plugin to display a
warning when a coding standard file has some comments that
can be erased when edited from the GUI.

An automatic test has also been added.

Change-Id: I1b96acc5b15531770277cac352883f152c30d856