S329-010: LSP-based Goto decl/body + ctrl-click
The 'goto declaration', 'goto body' actions and the hyper mode
click have been implemented based on the LSP.

It currently defaults to the old engine when the LSP is disabled
for the current buffer's language.

The support for dispatching calls has been improved: we now display
a contextual menu with all declarations/bodies proposals instead
of displaying a dialog.

Change-Id: Icac9d0c2d1b50b5bbe7e092575002eb0968f840a