S527-024: Scrollable tooltips for long contents
When the contents that are about to be displayed in tooltips
are too long, a scrolled window with a minimum size is now added
to the tooltip.

To allow scrolling them, editor tooltips are now placed just under
or above the hovered entity and tooltips are not hidden anymore if
the cursor hovers on them.

Change-Id: Ic093667af2739ecb4007951121fba5e4374facac