S328-050: Prevent unwanted rsync file duplication.
In remote mode, when setting the local directory via the GUI, the
directory path was returned without a separator at the end
(missing '/' at the end in Linux). This was the root of a nasty
issue: "rsync path/to/dir remote/path/" will copy "path/to/dir"
inside of "remote/path". What we want to do is "rsync path/to/dir/
remote/path/" to synchronize the directories.

Change-Id: Ia53443361e76f0af541c1da279efe8127cc21ae3