S606-042: Improve Search view behavior.
Allow the user to cycle through the result using the "Search"
action in the non-incremental mode (it was an incremental only
feature).
In incremental mode, properly preselect the text after a "Search"
action when the context was changed: we were resetting the search
on the same pattern.
Add a test.

Change-Id: Ibee11c117b8fd4458dc961f4da758141a11ecaec