S509-002 Use the exit request from gnat_server to exit gnat_server
Prior to this commit, the Python plugin would just kill the gnat_server
when the user click on "Exit Manual Proof". Now, the plugin send the
exit request so that gnat_server exits itself. But, if this does not work
and the user click twice on the "Exit Manual Proof" button, gnat_server
is killed as before.

* share/plug-ins/spark2014/itp_lib.py
(EXIT): string for the exit request
(parse_notif): Use exit_sent and send_request(0, EXIT) to activate the
death of gnat_server and the end of the plugin works.
(Tree_with_process.__init__): Initialize exit_sent
(Tree_with_process.start): Initialize exit_sent in case manual proof is
started twice in the same session.
(exit_by_gnat_server): Cover the case in which gnat_server dies for an
unexpected reason: the windows and the plugin work will still be removed
properly.
(send_request): Do a specific case for the Exit request (same as SAVE).
(Tree_with_process.exit): Include exit_sent in the logic so that we can
wait that the Exit request is processed by gnat_server before exiting.

Change-Id: I5b3edc4d5c8763918492bad6badda8434a844e46