S528-002 Disable sending new request to locate the next id
This would be launched at every notification of a new node. When the
number of nodes is over 9000 nodes (such as in cartesian_trees), it does
not work because there are as much request sent back to the server
(it makes GPS freeze).
This also partially disable the "feature" that would go to next goal when
one is proved but this feature was buggy and should be adressed in
S430-001.

* share/plug-ins/spark2014/itp_lib
(parse_notif): Remove the systematic call to get_next_id.

Change-Id: I56a895f4d18dc1bd5fa921b699fa96e71dd9fa54