S430-001 Reasonable jump to next vc
Now there should be a jump to next VC:
- when the current one becomes proved
- when a child of the current one becomes proved (going to another child
that is unproved)
- when a transformation is applied to get the focus on the transformation
node (from where cvc4 can be launched on this transformation children)

* share/plug-ins/spark2014/itp_lib.py
(TR_TYPE): node type of transformation
(parse_notif): on transformation, jump to the child. When something gets
proved send a request to the server to know where to go (note that this
should be safe as it is done only when is_init is set meaning that not a
lot of notification are coming).
(get_parent_node): Function for the hack to get a parent node id.
(node_jump_select): Correct a bug that would make this function fail when
the parent is outside of the focus (example: the current check is node 83
and the parent is the root node).
(select_function): Now update the selected_node which is memorized in the
plugin.
(get_next_id): Is now safe because it is conditioned on is_init. Also, we
can ask for the next id for the parent of the modified_id.

Change-Id: I940e1321fa9a16d28a3b1eb6ff054c8794680c73