S428-006 Select the first node after initialization of the proof tree
This allows selecting the first root node of the proof tree after it
was just initialized so that the user can enter commands directly.

* share/plug-ins/spark2014/itp_lib.py
(parse_message): Add argument abs_tree to be able to call
select_first_root.
(select_first_root): New function that select the first root of the tree.
The first root is one of the Why3 goal nodes that corresponds to the
check (a check can have several Why3 goal associated but not none).

Change-Id: Ie17f3ec972fa92e4063890e6b6cc4c94b7f37346