S702-017: Fix tooltip computation
There are 2 types of tooltips:
A - If not aligned with the tip area, the tooltip location
is determined by the cursor position when it was created.

B - If aligned with the tip area, the tooltip location is
determined by the tip area geometry and should not depend
on the cursor position.

Only for the tooltip of type B, we want to allow the navigation
between the tip area and the tooltip area.

Change-Id: Ia82fa96d5c91a27a2d86d36a870fd3e338e72644