S703-024: Always display LSP tooltips
We shoudl show the LSP-based tooltips even if we are not
hovering on entities, since we might want to display the
errors present on the hovered line.

Thus, we need to dislay the tooltip right after the query
by default and display it asynchronously only after sending
a LSP request (i.e: when hovering on an entity).

Change-Id: Ice35a9737476c2a9f978b95cfd6dd75879f820f7