S618-017 Allow space in path when using Manual proof
This solves a bug where adding spaces in path name would make manual proof
fail.

* share/plug-ins/spark2014.py
(start_ITP): Normalize the path of the .mlw file and add quotes around it
when using it in a command (Change that when GPS.Process support being
given several arguments like the subprocess of python).
Added some debugging

Change-Id: I3b8805907a389c7b0dcfd3085c73da9dd84c8fb7