S327-015 Improve proof tree display for manual proof
This adds a check column, reorder columns so that the name is first and
adds an ellipsize property so that the column "Status" and " " are always
visible even when wrapping occurs (the first column is wrapped).

* share/plug-ins/spark2014/itp_lib.py
(RED_CROSS): Unicode symbol for cross
(HEAVY_CHECK_MARK): Unicode symbol for check
(UPROVED): -> UPROVED_JSON or UPROVED_DIS. Differentiate when those are
used in the JSON format or in the display of the Treeview.
(create_check): Same as create_color. Convert information from the
treeview to a check/cross image.
(self.model): It now contains 3 new column: check, ellipsize_mode,
ellipsize. Note that the 2 last columns should be removed: another way of
doing the same thing should be found: we give a value to all the rows
whereas we could just set it for all of them at once.
(Tree.__init__): Reorder the column and add the attribute of ellipsization
(update_iter): Now update the check too.

Change-Id: If9eea411d04c3d7d7077e9a0e4c0662d1d34ea63