S711-057: Add defensive code in Log view
In particular disable the traces intercpetor when an exception
has been caught while catching traces.

Avoid intercating with the Log view when a trace has been caught:
display an error dialog to warn the user instead.

Register the interceptor when the Log view module is registered
instead of registering it at the very beginning: this ensures
that GPS will be properly initialized before trying to log the
traces in the Log view.

Change-Id: I8834a9a8d3aa23e8962346de1fc87bcfbe52e9f9