S403-032 Rework the GPS/LSP editor synchronization
Remove the need to maintain correspondance lists
between language servers and their managed documents,
between the documents and their server, and between
the documents and their source buffer.

This permits us to remove structures that are not
easy to maintain, and where de-synchronization can
lead to access to deallocated memory.

Add a high-level documentation in the body of GPS.LSP_Module.

Remove a number of subprograms that are no longer needed:
Associate, Dissociate, Dissociate_All, Text_Document,
Set_Server, Register, Unregister, Associated, Dissociated.

Add a primitive Get_Client, allowing to retrieve the
LSP_Client corresponding to a live server, if any.

Change-Id: Ifb6a72858eedf50975943000a6b90a8bf4a65586