S717-040 detect collisions in scenario variables
In corner cases where two scenario variables have the same name
(which looks like it can happen between a typed variable and an
untyped one), GPS cannot display the scenario view because of
this collision.

Detect this better, and show a message to the user to help them
sort out the collision.

Change-Id: I111d8dc89043d822953d684c8f585660fc279e6a