S814-037 Remove a hack in the reloading of files
We were emitting 'file_closed' and 'file_edited' around editor
reloads, even if the editor wasn't closed and reopened. The reason
for this was to keep the editor messages in the marker location.

Rather than this hack, introduce a hook 'file_reloaded' and
use this to replace the messages to their marks.

Change-Id: I40ec28eb7f77720efdccd18169ca593190a5269d