S910-029 Fix performance issue with insertion in locations view
Use a (naive and quadratic) algorithm for propagating weights
and backgrounds from leaf nodes to their parents.

This is not very efficient, but orders of magnitude faster
than the previous implementation which was relying on a
self-sorting tree model - unfortunately, that tree model
was re-sorting itself completely for each new message being
highlighted.

Change-Id: Id768f99930d50d97b84a5233e5822f002b8d8c74