QA12-026 itp: add interactive/rightclick menu for transformations name
This allows to have a menu in the GPS interface with most of available
transformations and provers. This menu is also displayed when
rightclicking on the tree.
This also prints a welcome message in the manual proof giving hints on how
to use it "help" "list-transforms" etc.

* share/plug-ins/spark2014/itp_lib.py
(MENU_NAME, WELCOME): New constants holding messages displayed
(transf_list): List of elements in the new submenu
(MenuITP): class for the definition of the submenu
(start_menu): go through the file generated_menus.json to add all
submenus.
This transformation list is generated in SPARK and installed: we actually
import the generated file there.
(kill_menu): go through ideal_transf_list to remove all submenus when
exiting ITP
(generation): Specific function used to generate the lambdas of the
callback of the menus.
(_on_view_button_press): connect the rightclick to the gtkmenu associated
to the new created menu (this makes the menu appear when rightclicking).
It is similar to a function with same name in lal python plugin.

* share/plug-ins/spark2014.py
(start_ITP): When calling start, add a new parameter to locate the install
directory inside itp_lib.py


Change-Id: I88e0f2ea18bbefe0dc959e4a224e531658fdd9bc