S912-037 Remove possibly fragile widget traversal
We are getting reports of exceptions of the form
   PROGRAM_ERROR : EXCEPTION_ACCESS_VIOLATION
occurring in C++ highlighting, when the engine traverses the
MDI child corresponding to an editor in search of the Gtk_Text_Buffer.

Attempt to circumvent this, or at least understand the issue better,
by removing the widget tree traversal, and offering direct access
to the underlying text buffer.

Change-Id: I09d10c0c8522bdb59039642bc59980366897c7b4