S925-044: Better handling of toolbars size preference
The toolbar buttons size preference is now better handler
in GNAT Studio: local toolbar buttons are now also impacted when
changing the preference.

The only icon buttons that can't be resized currently are the ones
in the Task widgets (HUD and Tasks view). This will probably be
possible when switching to a newer version of Gtk+.

Change-Id: Idc6fe8df768d24869d0371a7d6e9825df6b34a06