S919-029 Improve contextual Prove commands inside inlined code
Inside inlined code (typically for ghost code in inlined ghost procedures),
commands Prove Line and Prove Selected Region should not pass the switch
--limit-subp with the location of the enclosing subprogram, or else the
subprogram is analyzed in isolation, which defeats the purpose of having it
inlined.

Now fixed by checking first if the subprogram has or not a contract, which
is a condition for inlining. When in doubt (there is no contract) then the
switch --limit-subp is not passed.

* docs/users_guide/GPS/__init__.py
Fix missing newline leading to bad layout in generated doc.

* share/plug-ins/spark2014.py
(generic_on_analyze, on_prove_line): Add parameter force to be able to call
 build target from a Python test.
(build_limit_subp_string): Only pass --limit-subp switch when the enclosing
 subprogram may not be inlined.

Change-Id: Ia1bdbdd5a36eed560f4a67f4680b33b1114f1fbe