SA08-033 Remove some calls to fstat
In two functions that are called extremely often,
Language.Tree.Database.Get_Or_Create and Src_Editor_Module.Editors.Get,
move the disk access to the part of the function that's less likely
to be called.

This removes unneeded disk accesses that were occurring whenever
accessing an editor or refreshing the constructs tree.


Change-Id: Iefa331bde8ba46f51d956be3043a110879d05746