SA14-030: Commits view improvement
Clarify the Discard action: update tooltip and dialog message.
Refresh the VCS views after 'git --amend'.
Modify the behavior of "commit" to automatically stage all the
modified files if called when no file was staged.
Add a test.

Fix a magical import in vcs.py which was making dialogs visible
in the testsuite => change it by an explicit dependency.

Change-Id: Ic0fca8158a0b7be00a5f9c5ebc6b51f01982803f