S813-025: Popup menu at the ctrl-clicked location
We now display the menu that lists all the locations received
in a textDocument/definition response at the location that has
been ctrl-clicked, instead of always displaying the menu at
the pointer's location.

This is useful because textDocument/definition can take a long
time on dispatching calls: the user may have moved the pointer
somewhere else while waiting for the response.

Change-Id: I4c619cfd862787aa41ae32c67a004d641a53b3fc