SB22-005 Do not force the splash screen on top
When GNAT Studio takes a long time to load, it's interesting to see
the splash screen as a confirmation that something is loading, and
an annoyance for this window to be on top of all others. So, remove
the "always on top" setting.

Change-Id: Ida22db7493f17692303e9ef4c68d6d13bbcc9f54
(cherry picked from commit 72aa709a574db86fd64cfb15a63c5568090dd1d1)