SB22-011 Do not store scenario values when they are the default
When scenario variables have not been changed from their default
values in the project, do not store them in the properties.json
file, and do not pass them on the command line as part of -X
switches.

This allows for workflows where the project-defined default
value might change often, and where the common use case is
not to modify them manually.

Change-Id: If4b6b5a2d4b5e37800b721920d1437a9b089abec