R821-007: Add possibility to remove words on declarations
The Remove_Word_Command has been renamed to Remove_Words_Command
and now allows to remove several words with the same command.

This allows to also remove unexpected words in declarations when
explicitly needed.

This new feature is now used for the "could be 'in'..." parameter
and "'in' could be omitted" codefixes. 

Change-Id: Idf068b5ebf288535bc96d595f1bd156056bde98a