SB07-008: Fix tabulations preferences in editors
The preferences related to tabulations in editors were assuming
tab = 8 spaces which is wrong. Correctly use the indentation
level set by the user.

Add pango to modify the way the tabs are represented.

Disable C indent preference by default to have an uniform
behavior between version (the preference was doing nothing
in most of the case)

Add a test.

Change-Id: I33dc589bda5e8af0ec9a1ba7a4fa3a14cc6cc912