S919-029 Improve contextual Prove commands inside inlined code
Inside inlined code (typically for ghost code in inlined ghost
procedures), commands Prove Line, Prove Selected Region and Prove
Subprogram should not pass the switch --limit-subp with the location
of the enclosing subprogram, or else the subprogram is analyzed in
isolation, which defeats the purpose of having it inlined.

Now fixed by checking first if the subprogram has or not a contract,
either in aspect or pragma form, which is a condition for inlining.
Also check whether the subprogram is declared in a spec unit.
When in doubt (there is no contract and the subprogram is not declared
in a spec unit) then the switch --limit-subp is not passed. In such
a case, when analyzing the whole subprogram, an equivalent value for
--limit-region is passed so that the subprogram will be analyzed
whether or not it is inlined.

* docs/users_guide/GPS/__init__.py
Fix missing newline leading to bad layout in generated doc.

* share/plug-ins/spark2014.py
(generic_on_analyze, on_prove_line): Add parameter force to be able
 to call build target from a Python test.
(build_limit_subp_string): Add parameter whole_subp to distinguish
 when analysis of whole subprogram is meant.
(build_limit_subp_string): Only pass --limit-subp switch when the
 enclosing subprogram may not be inlined. Pass --limit-region instead
 when the subprogram may be inlined, and no region/line is already
 passed.

This reverts commit 605f667d95a64f187805bfdfdc07b94ed4423e19
which reverted commit 67bce2072882ee6be14a9399ee7b2abf518ea296.

Change-Id: I12f45ee1afcd1cfdb03263382cd123a020ea9232