T129-038 Phase out support for Gcov
Deactivate the UI for launching Gcov by default, preferring
GNATcoverage, unless the environment variable ENABLE_GCOV is set,
in which case honour the previous value of the preference.

Set ENABLE_GCOV in auto_set_path.py, so that the Community edition
continues showing the Gcov UI.

Implement this via the setting of an internal preference, so that
the tests can still be switched to test Gcov or GNATcov.

Change-Id: I7e08e85108d516fac3be4f71584f4843526513c9