T330-041 add an API to restart language servers
Minor refactor to reuse the code from GPS.LSP_Module to restart
language servers.

Improve the GPS.setenv API to store the user modifications in the
kernel environment.

Add an API to allow clients to retrieve the contents of the
kernel environment.

Use this API in the code that launches language servers to
set the environment accordingly.

Add documentation.
Add a behavioral test.

Change-Id: I7b45bbe4c4537a74952c593cd3a72db26bd433c6