T529-016: Secondaries locations in Spark messages
Separate the Spark messages and the compiler messages.
The first category must be handle by the Analysis report.
The second category is given directly to the Locations view.

Add Create_Tool_Message in the tool_parser API  and use it
to handle secondary locations in AnalysisTool messages.

Only show the Analysis Report when at least a Analysis Tool's
message has been created.

Change-Id: I80ea51731538e76da3aa1a579653af3d13743456