T828-033: Add a filter in the Call Stack view
To do so use a filtered Gtkada.Tree_View.Tree_View
Don't react to Selection_Changed: this event will be
triggered when cleaning or filtering the view.

Remove whitespaces in the Frame Number Column.
Remove +/-1 in the frame index computation, they starts at 0
not 1 (the python API was expected 0 as the first frame).

Add a test. 

Change-Id: I0b8cabd4f7b58ea7e6699ecdc8829e329f69350b