T908-026: Fix LSP server stopping code
Stopping a server from GS should always be considered as
intentional, independently from the fact that we should
reject incoming requests are not: this is needed to ensure
that we won't try to restart the server, thinking that it
has died unintentionally.

We can't add an automatic test for this, since many ada_language_server
instances can run at the same time during testing (parallelism). 

Change-Id: I3989911f7502695c1da67314bf8399ef51a447d5