T903-034: Add 32bit native targets in toolchains.py
To specify that the GNAT executables and the debugger should not
be prefixed by the target in this case. This avoids trying to spawn
'x86-windows-gdb' when using a 32bit Windows compiler for instance.

The code that spawns the debuger has been adapted: we should always
get the debugger executable from the toolchain when using GDB.

board_support.py has also been adapted to enable board debugging only
if the target is supported. 

Change-Id: Ib11c1e494c2425417475938a968db88453e17007