T921-005 Refine language server hook
Add parameters to the language_server_response_processed hook
so that clients can know which method, if any, was associated
to the response.

Use this to implement an utility wait_language_server that we
can use in testsuite drivers to wait for a response for the
right method: this is much safer than waiting for "any response"
and fiddle with timeouts in hope for the right one.

Use wait_language_server to stabilise some LSP tests.

(kernel.gpr): add Python to the sources list, since this contains
 hook.py

Rename the "data" parameter of language_server_response_processed
to "contents", since hooks.py generates code which uses a "Data"
variable which is sent to the callback instead of the real "data".

Change-Id: Ie5f96ff897ca8980784a19e407607e843f630430