TA21-011: Wrong location for OnTypeFormatting
We were looking for the trigger character from the end of the
selected range however it can be found multiple times inside
resulting into a buffer corruption => look for it from the start.
(The issue was hidden because of a missing -1 in the line
computation).

Furthermore, the LSP action was using an unitialized value
for Set which was introducing garbage.

Add a test.

Change-Id: I99689bd29fc805ceb4c91fda4d4c16dc3368d247
(cherry picked from commit 8e28e43d408a7ff25564e1511b3eaf9d84259404)