T514-007: Strip blanks after inserting a newline
When inserting a newline, we should try to strip the trailing blanks
of the current line after the insertion of the newline.

This is needed to avoid triggerring autocasing on the buffer without
the newline.

An automatic test has been added.

Change-Id: I1fdd49e969c6f988c57d89d4d95dc8470d719e37
(cherry picked from commit 2b5ddfdab101c3e5a42e54cab1aaaa3b819a99ff)