TB30-030: Add fallback mechanism to navigate in comments
When executing navigation requests (or ctrl-clicking) on a mixed-case
comment word, we now try to search for the first ocurrence of this word
in the current file to execute the textDocument/definition request on it.

This allows to ctrl-click on entities mentioned in comments that are
referenced in the current file.

A basic test has been added.

(no-precommit-check) 

Change-Id: I7bc2cd59c7aaf2dcb2130b5bbec13b9c19e7885d