U118-026 handle no global generation needed
The only case where GNATprove does not create a .gg file is when a
program unit does not need to generate globals. We now handle the lack
of a .gg file by informing the user that no global generation was
needed.

* show_globals.py
(edit_file): No .gg file is created when no globals were generated, so
 we now inform the user that no global generation was needed.

Change-Id: I3ff4e779a79186452b12a3296208980a633597cd