U126-015 Move the logic away from the plugin
GNATprove now deals with the cases where the subprogram is inlined for
proof, so that we don't need to perform the semantic quey p_decl_part which
is not guaranteed to work.

(on_prove_region): Replicate logic of on_prove_line.
(build_limit_subp_string): Remove special logic for inlined subprograms.
 Remove the need for the parameter whole_subp.

no-precommit-check

Change-Id: I30dd6e75c561b0891ffa1b5992b0f8b4a38792e7