U215-035: Add comment for phamtom preference.
We are creating a preference for each entry in preferences.xml.
It happens that a preference only exist for a version X of GS and
will be loaded by a version Y => we must do that to prevent
version Y from overwriting/removing the preference from
preferences.xml.
However the preference will have no effect in the version Y thus
add a comment explaining it.

Change-Id: I6ff36dc4b1a13e02a6cd142e080643062d18a347