U319-009 Restore display of counterexamples in SPARK plugin
The display of counterexamples was lost on commit a809b00a210 which fixed
a wrong reference to "file" with "ce_file" which was not in fact the file
name but the section for current/previous counterexample lines. Now
correctly fixed by passing the filename in parameter to the function that
expects it.

no-precommit-check

Change-Id: I0094b801a6ee121535d095f7c7b567fbe3009232