U318-041: Improve Signature Client GUI.
If defined use the ActiveParameter of the signature itself and not
the ActiveParameter of the ActiveSignature.

Allow Up/Down key to navigate in the signature list (special case
to give the focus in priority to the Completion window if opened).

Close the window when focusing out of the editor (like switching
workspace).

Fix the location of the window for multiscreen setup.

Lock the location of the window when clicking directly on the
arrows: this allow to easily click multiple time.

Prevent the window to be opened by non-interactive action (like
undo/redo)

Add tests.

Change-Id: I704b25966ce251bee6b7c81b974365ad34feab00