U416-020 Rework code for the CodeActions UI
Remove all the code in Refactoring.Code_Actions responsible
for creating a contextual menu for code actions: this is now
handled directly by Source_Editor_Buffer.Line_Information.

Move the action to create this menu to the Src_Editor_Module,
so it can benefit more than code actions.

Slight rework of this menu: align labels on the left rather
than centering them.

Adapt the test to take into account the removal of the
intermediary menu level.

Change-Id: I229c62fb420e3f778e6b07800af226bb73e7a8ef