S926-009: Prevent GUI freeze when setting git status
When git ls-tree returned a huge amount of file (like 100k)
the GUI will be frozen for 3-4s.
This is caused by:
- converting the list of 100k files into a set
and then doing a diff => directly construct a set inside the
asynchronous part. (2s)

- returning 100k files in the GS hook => do it step by step (5k)
inside a Glib.Timeout.

Add a hook to detect when the new timeout is done.

Change-Id: I6a5bf1a5d2e4494710cfdfd524ee680815f8f9c2