U315-006 clearcase.py: fix performance issue in getting status
The "cleartool ls -recurse" command might return a large amount of
lines corresponding to a large amount of filenames. These get
mapped to GPS.File instances - when calling the constructor with
a non-absolute path, these are resolved as sources, resulting in
calls to `stat` - a massive amount of these calls can be measurably
slow on ClearCase filesystems.

To avoid this potential performance issue, call "cleartool ls" with
the full path to the directory to list, which results in cleartool
returning absolute paths, therefore avoiding the calls to `stat`.

Change-Id: I81427998d2a369f26ef09e8eedbd5aebbcbc37a5