U409-032 update and clean "Regenerate Report" CodePeer menu
Take this occasion to replace a 'spin' to select analysis level with a
'combo' in order to be able to explicitly choose to not specify a level as
well as be able to specify level 0 (this is needed in case the default level
is changed by the user).

Also, remove the existence check of the Inspection_Info.xml file. This is
replaced by corresponding code in codeper_bridge.

Change-Id: Ieb6e057d54d779eb99a0ecdeb94a03a0e9fbec54