U112-028: revamp Analysis Report view
We now display the GNAThub metrics in the same tree view report
as the messages.

The Analysis Report is now divided in two separate tree views: one
that lists the project, dirs and entities, and another one that
lists the actual messages and metrics. Both trees are synchronized,
so that when the selection changes in one tree, we'll select the
same row in the other tree.

A filter panel has also been created for metrics, allowing the user
to show/hide particular metric columns, which can be useful if
the user displays a lot of metrics at the same time (e.g: when metrics
from several tools are displayed at the same time).

A basic test has been added and the other ones have been modified.

Change-Id: Ib584e758891deb63f0abcc29b5ab1802261be5c4