U708-014: Always display for codeActions
Add a flag in Line_Information_Record to control whether we should
show the popup when there is only one action and set it to True
for codeActions.

This is done to match VS Code's behavior, which is more natural, and
avoids executing an heavy codeAction, without knowing what it will
actually do, by clicking on the light-bulb icon.

Change-Id: Ie0950338b3236d437c8c6250da3ceb057c3cee41